Enable setting Horovod background thread affinity in general.
